Persist event for onChange event. (#176)
Since the onChange event prop is called in a setState callback, based on the nature of a Synthetic event it won't be persisted when passed to the this.props.onChange call at line 500. In order to correctly pass the event as third argument, it should be persisted before any async operation, otherwise it'll be nullified.